+++ src/constants.js
@@ -1,4 +1,4 @@
-export const VERSION = '2.3.1';
+export const VERSION = '2.4.0';
 
 export const SSH_COLORS = [
   'cyan',